backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
gspread==6.1.4
google-auth==2.37.0
pydantic-settings==2.7.0
